Conflicts:
	config/main.py

+++ config/main.py
@@ -34,6 +34,7 @@ from portconfig import get_child_ports
 
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help', '-?'])
 
+SONIC_GENERATED_SERVICE_PATH = '/etc/sonic/generated_services.conf'
 SONIC_CFGGEN_PATH = '/usr/local/bin/sonic-cfggen'
 SYSLOG_IDENTIFIER = "config"
 PORT_STR = "Ethernet"
@@ -44,6 +45,7 @@ BREAKOUT_CFG_FILE = get_port_config_file_name(hwsku, platform)
 
 VLAN_SUB_INTERFACE_SEPARATOR = '.'
 
+INIT_CFG_FILE = '/etc/sonic/init_cfg.json'
 
 # ========================== Syslog wrappers ==========================
 
@@ -521,6 +523,15 @@ def _get_platform():
                 return tokens[1].strip()
     return ''
 
+def _get_sonic_generated_services():
+    if not os.path.isfile(SONIC_GENERATED_SERVICE_PATH):
+        return None
+    generated_services_list = []
+    with open(SONIC_GENERATED_SERVICE_PATH) as generated_service_file:
+        for line in generated_service_file:
+            generated_services_list.append(line.rstrip('\n'))
+    return None if not generated_services_list else generated_services_list
+
 # Callback for confirmation prompt. Aborts if user enters "n"
 def _abort_if_false(ctx, param, value):
     if not value:
@@ -536,10 +547,18 @@ def _stop_services():
         'hostcfgd',
         'nat'
     ]
+    generated_services_list = _get_sonic_generated_services()
+
+    if generated_services_list is None:
+        log_error("Failed to get generated services")
+        return
+
     if asic_type == 'mellanox' and 'pmon' in services_to_stop:
         services_to_stop.remove('pmon')
 
     for service in services_to_stop:
+        if service + '.service' not in generated_services_list:
+            continue
         try:
             click.echo("Stopping service {} ...".format(service))
             run_command("systemctl stop {}".format(service))
@@ -567,7 +586,15 @@ def _reset_failed_services():
         'nat'
     ]
 
+    generated_services_list = _get_sonic_generated_services()
+
+    if generated_services_list is None:
+        log_error("Failed to get generated services")
+        return
+
     for service in services_to_reset:
+        if service + '.service' not in generated_services_list:
+            continue
         try:
             click.echo("Resetting failed status for service {} ...".format(service))
             run_command("systemctl reset-failed {}".format(service))
@@ -590,10 +617,18 @@ def _restart_services():
         'nat',
         'sflow',
     ]
+    generated_services_list = _get_sonic_generated_services()
+
+    if generated_services_list is None:
+        log_error("Failed to get generated services")
+        return
+
     if asic_type == 'mellanox' and 'pmon' in services_to_restart:
         services_to_restart.remove('pmon')
 
     for service in services_to_restart:
+        if service + '.service' not in generated_services_list:
+            continue
         try:
             click.echo("Restarting service {} ...".format(service))
             run_command("systemctl restart {}".format(service))
@@ -695,7 +730,11 @@ def reload(filename, yes, load_sysinfo, disable_validation):
         command = "{} -H -k {} --write-to-db".format(SONIC_CFGGEN_PATH, cfg_hwsku)
         run_command(command, display_cmd=True)
 
-    command = "{} -j {} --write-to-db".format(SONIC_CFGGEN_PATH, filename)
+    if os.path.isfile(INIT_CFG_FILE):
+        command = "{} -j {} -j {} --write-to-db".format(SONIC_CFGGEN_PATH, INIT_CFG_FILE, filename)
+    else:
+        command = "{} -j {} --write-to-db".format(SONIC_CFGGEN_PATH, filename)
+
     run_command(command, display_cmd=True)
     client.set(config_db.INIT_INDICATOR, 1)
 
@@ -709,7 +748,7 @@ def reload(filename, yes, load_sysinfo, disable_validation):
     _reset_failed_services()
     _restart_services()
 
-@config.command()
+@config.command("load_mgmt_config")
 @click.option('-y', '--yes', is_flag=True, callback=_abort_if_false,
                 expose_value=False, prompt='Reload mgmt config?')
 @click.argument('filename', default='/etc/sonic/device_desc.xml', type=click.Path(exists=True))
@@ -733,7 +772,7 @@ def load_mgmt_config(filename):
     run_command(command, display_cmd=True, ignore_error=True)
     click.echo("Please note loaded setting will be lost after system reboot. To preserve setting, run `config save`.")
 
-@config.command()
+@config.command("load_minigraph")
 @click.option('-y', '--yes', is_flag=True, callback=_abort_if_false,
                 expose_value=False, prompt='Reload config from minigraph?')
 def load_minigraph():
@@ -913,6 +952,91 @@ def remove(session_name):
     config_db.connect()
     config_db.set_entry("MIRROR_SESSION", session_name, None)
 
+#
+# 'pfcwd' group ('config pfcwd ...')
+#
+@config.group()
+def pfcwd():
+    """Configure pfc watchdog """
+    pass
+
+@pfcwd.command()
+@click.option('--action', '-a', type=click.Choice(['drop', 'forward', 'alert']))
+@click.option('--restoration-time', '-r', type=click.IntRange(100, 60000))
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+@click.argument('ports', nargs=-1)
+@click.argument('detection-time', type=click.IntRange(100, 5000))
+def start(action, restoration_time, ports, detection_time, verbose):
+    """
+    Start PFC watchdog on port(s). To config all ports, use all as input.
+
+    Example:
+        config pfcwd start --action drop ports all detection-time 400 --restoration-time 400
+    """
+    cmd = "pfcwd start"
+
+    if action:
+        cmd += " --action {}".format(action)
+
+    if ports:
+        ports = set(ports) - set(['ports', 'detection-time'])
+        cmd += " ports {}".format(' '.join(ports))
+
+    if detection_time:
+        cmd += " detection-time {}".format(detection_time)
+
+    if restoration_time:
+        cmd += " --restoration-time {}".format(restoration_time)
+
+    run_command(cmd, display_cmd=verbose)
+
+@pfcwd.command()
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+def stop(verbose):
+    """ Stop PFC watchdog """
+
+    cmd = "pfcwd stop"
+
+    run_command(cmd, display_cmd=verbose)
+
+@pfcwd.command()
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+@click.argument('poll_interval', type=click.IntRange(100, 3000))
+def interval(poll_interval, verbose):
+    """ Set PFC watchdog counter polling interval (ms) """
+
+    cmd = "pfcwd interval {}".format(poll_interval)
+
+    run_command(cmd, display_cmd=verbose)
+
+@pfcwd.command()
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+@click.argument('counter_poll', type=click.Choice(['enable', 'disable']))
+def counter_poll(counter_poll, verbose):
+    """ Enable/disable counter polling """
+
+    cmd = "pfcwd counter_poll {}".format(counter_poll)
+
+    run_command(cmd, display_cmd=verbose)
+
+@pfcwd.command()
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+@click.argument('big_red_switch', type=click.Choice(['enable', 'disable']))
+def big_red_switch(big_red_switch, verbose):
+    """ Enable/disable BIG_RED_SWITCH mode """
+
+    cmd = "pfcwd big_red_switch {}".format(big_red_switch)
+
+    run_command(cmd, display_cmd=verbose)
+
+@pfcwd.command()
+@click.option('--verbose', is_flag=True, help="Enable verbose output")
+def start_default(verbose):
+    """ Start PFC WD by default configurations  """
+
+    cmd = "pfcwd start_default"
+
+    run_command(cmd, display_cmd=verbose)
 
 #
 # 'qos' group ('config qos ...')
@@ -1086,6 +1210,7 @@ def add_vlan_member(ctx, vid, interface_name, untagged):
     db = ctx.obj['db']
     vlan_name = 'Vlan{}'.format(vid)
     vlan = db.get_entry('VLAN', vlan_name)
+    interface_table = db.get_table('INTERFACE')
 
     if get_interface_naming_mode() == "alias":
         interface_name = interface_alias_to_name(interface_name)
@@ -1105,6 +1230,10 @@ def add_vlan_member(ctx, vid, interface_name, untagged):
         else:
             ctx.fail("{} is already a member of {}".format(interface_name,
                                                         vlan_name))
+    for entry in interface_table:
+        if (interface_name == entry[0]):
+            ctx.fail("{} is a L3 interface!".format(interface_name))
+            
     members.append(interface_name)
     vlan['members'] = members
     db.set_entry('VLAN', vlan_name, vlan)
@@ -1304,7 +1433,8 @@ def add_vlan_dhcp_relay_destination(ctx, vid, dhcp_relay_destination_ip):
         return
     else:
         dhcp_relay_dests.append(dhcp_relay_destination_ip)
-        db.set_entry('VLAN', vlan_name, {"dhcp_servers":dhcp_relay_dests})
+        vlan['dhcp_servers'] = dhcp_relay_dests
+        db.set_entry('VLAN', vlan_name, vlan)
         click.echo("Added DHCP relay destination address {} to {}".format(dhcp_relay_destination_ip, vlan_name))
         try:
             click.echo("Restarting DHCP relay service...")
@@ -1329,7 +1459,11 @@ def del_vlan_dhcp_relay_destination(ctx, vid, dhcp_relay_destination_ip):
     dhcp_relay_dests = vlan.get('dhcp_servers', [])
     if dhcp_relay_destination_ip in dhcp_relay_dests:
         dhcp_relay_dests.remove(dhcp_relay_destination_ip)
-        db.set_entry('VLAN', vlan_name, {"dhcp_servers":dhcp_relay_dests})
+        if len(dhcp_relay_dests) == 0:
+            del vlan['dhcp_servers']
+        else:
+            vlan['dhcp_servers'] = dhcp_relay_dests
+        db.set_entry('VLAN', vlan_name, vlan)
         click.echo("Removed DHCP relay destination address {} from {}".format(dhcp_relay_destination_ip, vlan_name))
         try:
             click.echo("Restarting DHCP relay service...")
@@ -2253,6 +2387,48 @@ def platform():
 if asic_type == 'mellanox':
     platform.add_command(mlnx.mlnx)
 
+# 'firmware' subgroup ("config platform firmware ...")
+@platform.group()
+def firmware():
+    """Firmware configuration tasks"""
+    pass
+
+# 'install' subcommand ("config platform firmware install")
+@firmware.command(
+    context_settings=dict(
+        ignore_unknown_options=True,
+        allow_extra_args=True
+    ),
+    add_help_option=False
+)
+@click.argument('args', nargs=-1, type=click.UNPROCESSED)
+def install(args):
+    """Install platform firmware"""
+    cmd = "fwutil install {}".format(" ".join(args))
+
+    try:
+        subprocess.check_call(cmd, shell=True)
+    except subprocess.CalledProcessError as e:
+        sys.exit(e.returncode)
+
+# 'update' subcommand ("config platform firmware update")
+@firmware.command(
+    context_settings=dict(
+        ignore_unknown_options=True,
+        allow_extra_args=True
+    ),
+    add_help_option=False
+)
+@click.argument('args', nargs=-1, type=click.UNPROCESSED)
+def update(args):
+    """Update platform firmware"""
+    cmd = "fwutil update {}".format(" ".join(args))
+
+    try:
+        subprocess.check_call(cmd, shell=True)
+    except subprocess.CalledProcessError as e:
+        sys.exit(e.returncode)
+
 #
 # 'watermark' group ("show watermark telemetry interval")
 #
@@ -2731,6 +2907,42 @@ def feature_status(name, state):
         return
 
     config_db.mod_entry('FEATURE', name, {'status': state})
+
+#
+# 'container' group ('config container ...')
+#
+@config.group(name='container', invoke_without_command=False)
+def container():
+    """Modify configuration of containers"""
+    pass
+
+#
+# 'feature' group ('config container feature ...')
+#
+@container.group(name='feature', invoke_without_command=False)
+def feature():
+    """Modify configuration of container features"""
+    pass
+
+#
+# 'autorestart' subcommand ('config container feature autorestart ...')
+#
+@feature.command(name='autorestart', short_help="Configure the status of autorestart feature for specific container")
+@click.argument('container_name', metavar='<container_name>', required=True)
+@click.argument('autorestart_status', metavar='<autorestart_status>', required=True, type=click.Choice(["enabled", "disabled"]))
+def autorestart(container_name, autorestart_status):
+    config_db = ConfigDBConnector()
+    config_db.connect()
+    container_feature_table = config_db.get_table('CONTAINER_FEATURE')
+    if not container_feature_table:
+        click.echo("Unable to retrieve container feature table from Config DB.")
+        return
+
+    if not container_feature_table.has_key(container_name):
+        click.echo("Unable to retrieve features for container '{}'".format(container_name))
+        return
+
+    config_db.mod_entry('CONTAINER_FEATURE', container_name, {'auto_restart': autorestart_status})
 
 if __name__ == '__main__':
     config()
